Update routes and API endpoint; Improve UI and accessibility elements
In this commit, the navigation and API endpoints are updated and the UI is enhanced. The 'Planet' route was removed from the app, and the API endpoint for login and registration was updated to target the production server instead of localhost. Additionally, Card and SearchAppBar components were updated for better accessibility and user interface experience. On top of that, the unwanted backdrop prop was removed from Drawer component for a cleaner UI. Finally, some console.log statements were added in 'contextPage.tsx' for debugging purposes.