Update getGeocode input and adjust display of distance
The getGeocode function now takes "solstad terrasse 30" as an input instead of "asker trekanten". The calculated distance is converted from meters to kilometers and is displayed as such. In addition, some minor code clean-up is performed, such as the refactoring of the LocationDistance component from an arrow function to a regular function.