Move LocationDistance component to Filter.tsx
The LocationDistance component, that was previously situated in the App.tsx, has been moved to the newly created Filter.tsx file. Furthermore, a minor correction in the fetchLocationsAndComputeDistance async function - removing an accidental whitespace at the beginning of "solstad terrasse 30". Also, a Filter component has been added to the Home.tsx for future filtering functionalities.