Reformat HTML, JS and CSS code
The primary emphasis of this commit was to improve the overall readability and maintainability of the system's HTML, JavaScript and CSS files. Changes include correcting improper nesting and multiliner formatting, fixing syntax inconsistencies and removing unnecessary imports and variables. These modifications do not modify system behavior, rather they aim to make the code more understandable and easier to work with.