Refactor styles in Filter component
Adjusted a number of style properties in the Filter file to enhance UI appearance. The refactor includes a border style addition, an implementation of Display and Transition properties for smoother transitions, as well as adjustments to the Box styling. These changes aim to encompass an improved and more user-friendly interaction.